Add files via upload
Instructions  
Use the following api to get the data required for this assignment: https://openweathermap.
org/current.
You need to create an account to get an API key. They have a free tier so no need to buy 
any premium plan.
Tasks  
1.  Create a function that accepts any number of countries and creates a graph showing 
the current temperature for each provided country.
2.  Show in a map the current temperature for the following countries Sweden, Norwary, 
Denmark, Spain, and Panama. The size of the marker is associated to the pressure.